pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wxcadm"
version = "4.6.0"
authors = [
  { name="Trey Hilyard", email="kctrey@gmail.com" },
]
dependencies = [
    "requests>=2.31.0",
    "requests-toolbelt>=1.0.0",
    "srvlookup>=2.0.0",
    "xmltodict>=0.12.0",
    "pyhumps==3.8.0",
    "dataclasses-json>=0.6.4"
]
description = "A Python 3 library for Webex Calling administrators"
license = "GPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.setuptools]
packages = ["wxcadm"]

[project.urls]
Homepage = "https://github.com/kctrey/wxcadm"
Issues = "https://github.com/kctrey/wxcadm/issues"
Documentation = "https://wxcadm.readthedocs.io"
Changelog = "https://wxcadm.readthedocs.io/en/latest/whatsnew.html"

[project.optional-dependencies]
meraki = [
    "meraki>=1.30.0"
]